+++ resources/js/app.js
@@ -1,9 +1,11 @@
 // import './bootstrap';
 
 import Alpine from 'alpinejs'
-import Htmx from 'htmx.org'
+import morph from '@alpinejs/morph'
+import htmx from 'htmx.org'
 
+window.htmx = htmx
 window.Alpine = Alpine
-window.Htmx = Htmx
+Alpine.plugin(morph)
 
 Alpine.start()