+++ vite.config.js
@@ -1,8 +1,13 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
+import inject from "@rollup/plugin-inject";
 
 export default defineConfig({
     plugins: [
+        inject({
+            // Add HTMX Extensions here
+            'htmx-alpine-morph': 'htmx.org/dist/ext/alpine-morph.js'
+        }),
         laravel({
             input: ['resources/css/app.css', 'resources/js/app.js'],
             refresh: true,