+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `@rollup/plugin-inject` to devDependencies.
+- Added `@alpinejs/morph` to dependencies.
+- Added `htmx-alpine-morph` extension to `vite.config.js` using `@rollup/plugin-inject`.
+
 ## [1.0.0] - 2023-05-23
 
 ### Added